% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankData.R
\docType{data}
\name{bank}
\alias{bank}
\title{Arrival time and service time of customers from a bank}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{X}{customer number, integer}
\item{arrival_time}{customer's arrival time, in seconds}
\item{service_time}{customer's service time, in seconds}
...
}
}
\usage{
bank
}
\description{
A dataset containing the arrival time and service time of 100 customers
}
\keyword{datasets}
