% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Customer queue simulation}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{double; This is the time customer will arrive}

\item{ServiceTimes}{double; This is how long the service is going to be}

\item{NumServers}{double; This is the server they will be using}
}
\value{
A dataframe that show the customer transition times
}
\description{
Feed customers through a multiserver queue system to determine each customer's transition times.
}
